+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 
 import App from './App'
-import reportWebVitals from './reportWebVitals'
+import reportWebVitals from './utils/reportWebVitals'
 
 ReactDOM.render(
   <React.StrictMode>
